requirements.txt
simplejson==3.6.5
mysql-connector-python==2.0.3
requests==2.6.0
iso8601==0.1.10